fix: Blogger API位置情報送信の名前空間エラーを修正
- uploader.py: 名前空間を正しいURLに修正 (http://www.blogger.com/atom/ns#)
- uploader.py: 位置情報抽出ロジックを<blogger:location>タグに対応
- uploader.py: デバッグ出力を追加（送信データの可視化）
- convert_atom.py: 不要なタグ削除（<georss:point>, <georss:featurename>）
- config.py/ini: MAX_POSTS_PER_RUN=50に設定

根本原因: Blogger API名前空間の誤り（2018 → atom/ns#）により
位置情報フィールドが読み取れず、送信されていなかった。
修正後、全4件のテスト投稿で位置情報が正常に保存されることを確認。

+++ config.py
@@ -103,7 +103,7 @@ MEDIA_MANAGER_FILE = get_config('UPLOADER', 'MEDIA_MANAGER_FILE', DEFAULTS.get('
 LOG_FILE = get_config('UPLOADER', 'LOG_FILE', DEFAULTS.get('LOG_FILE'))
 SCOPES_STR = get_config('UPLOADER', 'SCOPES', DEFAULTS.get('SCOPES'))
 SCOPES = [s.strip() for s in SCOPES_STR.split(',')]
-DELAY_SECONDS = int(get_config('UPLOADER', 'DELAY_SECONDS', DEFAULTS.get('DELAY_SECONDS')))
+DELAY_SECONDS = float(get_config('UPLOADER', 'DELAY_SECONDS', DEFAULTS.get('DELAY_SECONDS')))  # floatに変更
 MAX_POSTS_PER_RUN = int(get_config('UPLOADER', 'MAX_POSTS_PER_RUN', DEFAULTS.get('MAX_POSTS_PER_RUN')))
 
 def open_file_with_default_app(filepath):
